# [Bugfix][VIP] Fix MLA chunked prefill performance with mini overhead